fix #21 - toKstTest 제거 및 forecast 온도 반올림 적용

+++ WeatherApp/Model/ForecastWeatherResult.swift
@@ -35,8 +35,7 @@ struct ForecastWeather: Codable {
 
 struct DailyWeather {
     let day: String   // "yyyy-MM-dd"
-    let minTemp: Double
-    let maxTemp: Double
+    let minTemp: Int
+    let maxTemp: Int
     let weatherIcon: String
 }
-
